Eliminate transformation pattern using WILL in line 9182
The "transformations" facility apparently assumes that all verbs have been previously entered as primaries, and in fact since that code is only a prototype that might have been why the auxiliary verbs WILL and WOULD were included in the the dictionary in first place in the rather implausible location. So in their absence, this pattern causes a key error.